% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_in_file.R
\name{mod_in_file_server}
\alias{mod_in_file_server}
\title{in_file Server Function}
\usage{
mod_in_file_server(
  input,
  output,
  session,
  in_ras,
  HTboxI,
  file_path,
  clear_map,
  change_ht,
  input_box,
  feet,
  values,
  switch_fil,
  met,
  lab_sel
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{in_ras}{A reactiveValues that contains numerous rasters}

\item{HTboxI}{A reactive that holds the change height.}

\item{file_path}{A reactive that let's the function know that it can start running (e.g. after a file path has been recognized).}

\item{clear_map}{This is a reactive input that will clear everything back to the beginning. Some checkboxs will not
be set back to original (feet/meters) but height and aoi will.}

\item{change_ht}{A reactive that initiates the button to change height.}

\item{input_box}{A reactiveValues that when the app is cleared the heights will go back to the original values.}

\item{feet}{A reactive that initiates the observeEvent to look at whether to use feet or meters when converting raster.}

\item{values}{A reactiveValues that stores the sf information when cropping leaflet map. This makes it possible to crop but also let's
the app know what metric to use (feet/meters).}

\item{switch_fil}{A reactive that actually changes back/forth from feet/meters}

\item{met}{A reactive that initiates the observeEvent to look at whether to use feet or meters when converting raster.}

\item{lab_sel}{A reactive that initiates the option for labeling in ft or meters or z.}
}
\description{
in_file Server Function
}
