% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_panel_3d.R
\name{mod_panel_3d_server}
\alias{mod_panel_3d_server}
\title{panel_3d Server Function}
\usage{
mod_panel_3d_server(
  input,
  output,
  session,
  in_ras,
  clear_map,
  smooth3d,
  sws,
  values
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{in_ras}{A reactiveValues that contains numerous rasters}

\item{clear_map}{This is a reactive input that will clear everything back to the beginning. Some checkboxs will not
be set back to original (feet/meters) but height and aoi will.}

\item{smooth3d}{Let's the server know whether to switch to smoothing or not. Reactive.}

\item{sws}{Reactive that determines the window size of smoothing.}

\item{values}{A reactiveValues that stores the sf information when cropping leaflet map. This makes it possible to crop but also let's
the app know what metric to use (feet/meters).}
}
\description{
panel_3d Server Function
}
