% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_panel_3d.R
\name{mod_panel_3d_ui}
\alias{mod_panel_3d_ui}
\title{panel_3d UI Function}
\usage{
mod_panel_3d_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A shiny Module that generates the 3d model of the input raster.
}
