% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_panel_leaflet.R
\name{mod_panel_leaflet_ui}
\alias{mod_panel_leaflet_ui}
\title{panel_leaflet UI Function}
\usage{
mod_panel_leaflet_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A shiny Module that controls the leaflet map panel. This function has the ability to crop, draw and change overlays.
}
