% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_panel_plot_map.R
\name{mod_panel_plot_map_server}
\alias{mod_panel_plot_map_server}
\title{panel_plot_map Server Function}
\usage{
mod_panel_plot_map_server(input, output, session, in_ras)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{in_ras}{A reactiveValues that contains numerous rasters.}
}
\description{
panel_plot_map Server Function
}
