% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_panel_plot_map.R
\name{mod_panel_plot_map_ui}
\alias{mod_panel_plot_map_ui}
\title{panel_plot_map UI Function}
\usage{
mod_panel_plot_map_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A shiny Module defaults to a basic raster image if there is no projection on the fileInput raster.
}
