% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_panel_stat_plot.R
\name{mod_panel_stat_plot_ui}
\alias{mod_panel_stat_plot_ui}
\title{panel_stat_plot UI Function}
\usage{
mod_panel_stat_plot_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A shiny Module that controls the plotting section of the panels. This let's the user switch between plots and also metric labels (feet/meters). It also is reactive
to the cropping feature in leaflet.
}
