% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_panel_stats.R
\name{mod_panel_stats_server}
\alias{mod_panel_stats_server}
\title{panel_stats Server Function}
\usage{
mod_panel_stats_server(input, output, session, in_ras, clear_map, values)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{in_ras}{A reactiveValues that contains numerous rasters.}

\item{clear_map}{This is a reactive input that will clear everything back to the beginning. Some checkboxs will not
be set back to original (feet/meters) but height and aoi will.}

\item{values}{A reactiveValues that stores the sf information when cropping leaflet map. This makes it possible to crop but also let's
the app know what metric to use (feet/meters).}
}
\description{
panel_stats Server Function
}
