% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_panel_stats.R
\name{mod_panel_stats_ui}
\alias{mod_panel_stats_ui}
\title{panel_stats UI Function}
\usage{
mod_panel_stats_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A shiny Module that controls the stats section of the panels. This let's the user switch metric labels (feet/meters) but is also reactive
to the cropping feature in leaflet.
}
